notification-system/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
fastapi==0.104.0
h11==0.14.0
idna==3.4
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
redis==5.0.1
sniffio==1.3.0
starlette==0.27.0
types-pyopenssl==23.3.0.0
types-redis==4.6.0.7
typing-extensions==4.8.0
uvicorn==0.23.2
supervisor
boto3
locust
aiohttp
aioapns
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-exporter-otlp==1.21.0
prometheus-client==0.17.1
grafana-dashboard-builder==0.6.0a1
